fix(solc-comparison): fail on unknown isoltest settings, and keep EVM target parsing out of infra_utils
