+++ crates/solidity-v2/testing/solc/Cargo.toml
@@ -10,6 +10,7 @@ path = "src/lib.rs"
 
 [dependencies]
 anyhow = { workspace = true }
+Inflector = { workspace = true }
 infra_utils = { workspace = true }
 rayon = { workspace = true }
 semver = { workspace = true }
