+++ crates/solidity-v2/testing/solc/src/test_case.rs
@@ -1,9 +1,7 @@
 use std::path::Path;
 
 use anyhow::{Context, Result, bail};
-use infra_utils::solc::parse_evm_target_name;
 use slang_solidity_v2_common::collections::OrderedMap;
-use slang_solidity_v2_common::evm_targets::EvmTarget;
 
 /// A single semantic test, parsed from the `isoltest` file format into the set
 /// of in-memory source files it comprises plus the compiler settings that
@@ -32,6 +30,9 @@ pub struct IsolTestCase {
     /// The raw `EVMVersion` setting (e.g. `>=byzantium`), if present. Resolved
     /// to a concrete [`EvmTarget`] by the runner, which knows the language
     /// version being used (see [`resolve_evm_target`]).
+    ///
+    /// [`EvmTarget`]: slang_solidity_v2_common::evm_targets::EvmTarget
+    /// [`resolve_evm_target`]: crate::evm_target::resolve_evm_target
     pub evm_version: Option<String>,
 }
 
@@ -43,7 +44,11 @@ impl IsolTestCase {
             .with_context(|| format!("Failed to read test file: {test_path:?}"))?;
 
         let (source_region, settings) = split_trailer(&contents);
-        let evm_version = setting_value(&settings, "EVMVersion");
+
+        check_settings_are_known(&settings)
+            .with_context(|| format!("Unsupported settings in test file: {test_path:?}"))?;
+
+        let evm_version = setting_value(&settings, EVM_VERSION_SETTING);
 
         // The file id used for the implicit (undelimited) source is the test's
         // own file name, matching how `solc` names a single-file input.
@@ -116,6 +121,46 @@ fn setting_value<'a>(settings: &[(&'a str, &'a str)], key: &str) -> Option<&'a s
         .iter()
         .find(|(name, _)| *name == key)
         .map(|(_, value)| *value)
+}
+
+/// The one setting that changes how we analyze a test.
+const EVM_VERSION_SETTING: &str = "EVMVersion";
+
+/// The settings we knowingly ignore.
+const ALLOWED_SETTINGS: &[&str] = &[
+    EVM_VERSION_SETTING,
+    // All the other settings are ignored
+    // ----------------------------------
+    // Which code generation pipeline(s) to run the test through, and in which
+    // bytecode format to emit the result.
+    "compileViaYul",
+    "compileViaSSACFG",
+    "compileToEwasm",
+    "bytecodeFormat",
+    // Code generation options, likewise.
+    "revertStrings",
+    // Restricts which ABI coder the test runs under.
+    "ABIEncoderV1Only",
+    // The same, for the experimental language mode: the source carries the
+    // `experimental` pragma, so we compile it exactly as written.
+    "experimental",
+    // Lets the runtime expectations call functions that don't exist, which only
+    // concerns the expectations — and we ignore those entirely.
+    "allowNonExistingFunctions",
+];
+
+/// Rejects any setting that isn't one of the two above.
+fn check_settings_are_known(settings: &[(&str, &str)]) -> Result<()> {
+    for (key, _) in settings {
+        if !ALLOWED_SETTINGS.contains(key) {
+            bail!(
+                "unrecognized isoltest setting {key:?}. If it cannot affect whether the test \
+                 compiles, add it to 'ALLOWED_SETTINGS'."
+            );
+        }
+    }
+
+    Ok(())
 }
 
 /// Parses the source region into named files, following `Source` and
@@ -228,103 +273,6 @@ fn load_external_source(
     push_source(files, source_name, content)
 }
 
-/// `isoltest`'s placeholder for an EVM version that has not been released yet,
-/// written in place of a target name (e.g. `EVMVersion: =@future`).
-pub const FUTURE_EVM_VERSION: &str = "@future";
-
-/// A successfully understood `EVMVersion` setting.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum ParsedTarget {
-    /// The test asks for [`FUTURE_EVM_VERSION`], an EVM version that hasn't
-    /// been released yet, so there is no target to analyze it at.
-    FutureSpec,
-    /// The concrete target to analyze the test at.
-    Target(EvmTarget),
-}
-
-/// Resolves the `EVMVersion` setting (e.g. `>=byzantium`, `<constantinople`,
-/// `istanbul`) to a single concrete [`EvmTarget`] to analyze the test at.
-///
-/// `default` is the EVM target `solc` of the language version being tested would
-/// pick on its own (the runner passes the version's default). We prefer it
-/// whenever it satisfies the constraint, so a lower-bound-only setting like
-/// `>=byzantium` stays at the version's real default rather than jumping to the
-/// newest target ever — which is a combination that `solc` of that version never
-/// ran, and would spuriously enable target-gated builtins. Only when the default
-/// doesn't satisfy the constraint (e.g. a hard `<constantinople` upper bound on a
-/// newer default) do we pick the *nearest* satisfying target to the default.
-///
-/// When no setting is present, `default` is used.
-///
-/// [`ParsedTarget::FutureSpec`] means the test asks for an EVM version `slang`
-/// has no target for at all. For now this only covers the [`FUTURE_EVM_VERSION`]
-/// isoltest marker.
-///
-/// Anything else *is* an error: silently falling back on a setting we failed to
-/// understand would analyze the test at the wrong target and quietly bake the
-/// result into the baseline, so a setting `solc` accepts but we don't is treated
-/// as a gap in *this* code.
-pub fn resolve_evm_target(setting: Option<&str>, default: EvmTarget) -> Result<ParsedTarget> {
-    let Some(setting) = setting else {
-        return Ok(ParsedTarget::Target(default));
-    };
-
-    let setting = setting.trim();
-    let (op, name) = if let Some(name) = setting.strip_prefix(">=") {
-        (">=", name)
-    } else if let Some(name) = setting.strip_prefix("<=") {
-        ("<=", name)
-    } else if let Some(name) = setting.strip_prefix('>') {
-        (">", name)
-    } else if let Some(name) = setting.strip_prefix('<') {
-        ("<", name)
-    } else if let Some(name) = setting.strip_prefix('=') {
-        ("=", name)
-    } else {
-        ("=", setting)
-    };
-
-    let name = name.trim();
-
-    // `isoltest`'s placeholder for the next, not-yet-released EVM version.
-    if name.eq_ignore_ascii_case(FUTURE_EVM_VERSION) {
-        return Ok(ParsedTarget::FutureSpec);
-    }
-
-    let bound = parse_evm_target_name(name)
-        .with_context(|| format!("Unrecognized EVM target in 'EVMVersion: {setting}'."))?;
-
-    let satisfies = |target: EvmTarget| match op {
-        ">=" => target >= bound,
-        ">" => target > bound,
-        "<=" => target <= bound,
-        "<" => target < bound,
-        _ => target == bound,
-    };
-
-    // Prefer the version's own default target when it already satisfies the
-    // constraint.
-    if satisfies(default) {
-        return Ok(ParsedTarget::Target(default));
-    }
-
-    // Otherwise pick the supported target satisfying the constraint that sits
-    // closest to the default (the nearest above for an unmet lower bound, the
-    // nearest below for an unmet upper bound).
-    let index_of = |target: EvmTarget| EvmTarget::ALL.iter().position(|t| *t == target);
-    let default_index = index_of(default);
-    EvmTarget::ALL
-        .iter()
-        .copied()
-        .filter(|target| satisfies(*target))
-        .min_by_key(|target| match (index_of(*target), default_index) {
-            (Some(i), Some(d)) => i.abs_diff(d),
-            _ => usize::MAX,
-        })
-        .map(ParsedTarget::Target)
-        .with_context(|| format!("No supported EVM target satisfies 'EVMVersion: {setting}'."))
-}
-
 /// Yields `(byte offset, line)` for each line in `contents`.
 fn line_offsets(contents: &str) -> impl Iterator<Item = (usize, &str)> {
     contents.lines().scan(0usize, |offset, line| {
@@ -362,85 +310,32 @@ contract C {}
         assert_eq!(setting_value(&settings, "missing"), None);
     }
 
+    /// A setting we've never seen has to be looked at, rather than ignored on
+    /// the assumption that it doesn't matter.
+    #[test]
+    fn rejects_settings_it_does_not_know() {
+        let check = |line: &str| {
+            let contents = format!("contract C {{}}\n// ====\n{line}\n");
+            let (_, settings) = split_trailer(&contents);
+
+            check_settings_are_known(&settings)
+        };
+
+        assert!(check("// EVMVersion: >=byzantium").is_ok());
+        assert!(check("// compileViaYul: also").is_ok());
+        assert!(check("// nonesuch: true").is_err());
+
+        // Casing is part of the name: a near-miss is still a setting we've
+        // never looked at.
+        assert!(check("// evmVersion: >=byzantium").is_err());
+    }
+
     #[test]
     fn source_without_trailer_is_all_source() {
         let contents = "pragma solidity >=0.4.0;\ncontract C {}\n";
         let (source, settings) = split_trailer(contents);
         assert_eq!(source, contents);
         assert!(settings.is_empty());
-    }
-
-    #[test]
-    fn resolves_evm_version_constraints() {
-        let resolve = |setting, default| match resolve_evm_target(setting, default).unwrap() {
-            ParsedTarget::Target(target) => target,
-            ParsedTarget::FutureSpec => panic!("'{setting:?}' should resolve to a concrete target"),
-        };
-
-        // No setting: the version's default target is used as-is.
-        assert_eq!(resolve(None, EvmTarget::Istanbul), EvmTarget::Istanbul);
-
-        // A lower bound the default already satisfies keeps the *version's*
-        // default rather than jumping to the newest target ever — the crux of
-        // the version-aware fix (a `>=byzantium` test at an Istanbul-default
-        // version runs at Istanbul, not Amsterdam).
-        assert_eq!(
-            resolve(Some(">=byzantium"), EvmTarget::Istanbul),
-            EvmTarget::Istanbul
-        );
-
-        // If the default is too old for the lower bound, pick the nearest
-        // satisfying target above it.
-        assert_eq!(
-            resolve(Some(">=cancun"), EvmTarget::Istanbul),
-            EvmTarget::Cancun
-        );
-
-        // Upper bounds the default violates resolve to the nearest satisfying
-        // target below it.
-        assert_eq!(
-            resolve(Some("<constantinople"), EvmTarget::Osaka),
-            EvmTarget::Byzantium
-        );
-        assert_eq!(
-            resolve(Some("<=constantinople"), EvmTarget::Osaka),
-            EvmTarget::Constantinople
-        );
-
-        // Exact constraints resolve to that target regardless of the default.
-        assert_eq!(
-            resolve(Some("=istanbul"), EvmTarget::Osaka),
-            EvmTarget::Istanbul
-        );
-        assert_eq!(
-            resolve(Some("istanbul"), EvmTarget::Osaka),
-            EvmTarget::Istanbul
-        );
-    }
-
-    #[test]
-    fn yields_no_target_for_an_unreleased_evm_version() {
-        // `@future` names an EVM version that doesn't exist yet, so there is
-        // nothing to analyze the test at. That's reported as `FutureSpec` rather
-        // than an error, because it says something about the test rather than
-        // about this code — and the runner turns it into a failing test.
-        for setting in ["=@future", "@future", ">=@future", "=@FUTURE"] {
-            assert_eq!(
-                resolve_evm_target(Some(setting), EvmTarget::Istanbul).unwrap(),
-                ParsedTarget::FutureSpec,
-                "'{setting}' should resolve to no target"
-            );
-        }
-    }
-
-    #[test]
-    fn rejects_evm_version_settings_it_cannot_honor() {
-        // A name we don't know is a gap in *our* table: resolving it to the
-        // default would analyze the test at a target `solc` never used.
-        assert!(resolve_evm_target(Some("nonsense"), EvmTarget::Osaka).is_err());
-
-        // Likewise a well-formed constraint no supported target satisfies.
-        assert!(resolve_evm_target(Some(">amsterdam"), EvmTarget::Osaka).is_err());
     }
 
     #[test]
